get list off any user by giving userid

+++ src/main/scala/com/anilist/controllers/anicontroller.scala
@@ -5,7 +5,7 @@ import com.anilist.models._
 object anicontroller {
 
   def getUserAnimelist(userid: String): String = {
-    AnimeData.getAnimeList(userid)
+    AnimeData.getAnimeList(userid.toInt)
   }
-
 }
+
